修复Altera SPI IP-Core 的转义字符导致的符号不同步问题

+++ DigiC-PC/inc/FTDUtil.h
@@ -10,9 +10,7 @@ public:
 	int lsDev();
 	int lsSPIChannel();
 	void SPIWriteByte(uint8 slaveAddress, uint8 address, uint32 data);
-	void SPIWriteByteArray(uint8 * Data,int);
-
-
+	void SPIWriteByteArray(uint8 * Data,long long int);
 	ChannelConfig mSPIChannelConf;
 	FT_HANDLE mHandle;
 private :
